correct types.

+++ examples/memcopy/cmd/test-memcopy/main.go
@@ -36,7 +36,7 @@ func main() {
 
 		krnl.SetMemoryArg(0, inputBuff)
 		krnl.SetMemoryArg(1, outputBuff)
-		krnl.SetArg(2, len(input))
+		krnl.SetArg(2, uint32(len(input)))
 
 		krnl.Run(1, 1, 1)
